# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Higgsup KPI APIS",
    "title": "Higgsup KPI APIS",
    "version": "2.0"
  },
  "host": "${host}",
  "basePath": "/",
...
  "security": [
    {
      "Authorization": []
    }
  ],
  "securityDefinitions": {
    "Authorization": {
      "description": "Get token at API with URI '/api/login' in login-controller ",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/events/support": {
      "post": {
        "tags": [
          "event-controller"
        ],
        "summary": "createSupport",
        "operationId": "createSupportUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "supportDTO",
            "description": "supportDTO",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "beginDate",
                "group",
                "additionalConfig"
              ],
              "properties": {
                "beginDate": {
                  "type": "string"
                },
                "group": {
                  "type": "object",
                  "required": [
                    "id"
                  ],
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                },
                "additionalConfig": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "name",
                      "quantity"
                    ],
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "quantity": {
                        "type": "number"
                      }
                    }
                  }
                }
              },
              "example": {
                "beginDate": "10-09-2018 11:00",
                "group": {
                  "id": 2
                },
                "additionalConfig": [
                  {
                    "name": "cleanUpPoint",
                    "quantity": 1
                  }
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "status_code : 200. Meaning: Thành công \n status_code : 901 \n Message: start date cannot null. Meaning: ngày bắt đầu không được null. \n Message: group cannot null. Meaning: group không thể rỗng \n Message: group ID can not be null. Meaning: ID group không được rỗng \n Message: group ID not is support. Meaning: ID group không phải là support \n Message: list support can not empty. Meaning: danh sách support không được rỗng \n Message: name at index {index} can not null. Meaning: tên tại 1 không thể rỗng \n Message: name support at index {index} not incorrect. Meaning: tên support tại 1 không chính xác \n Message: quantity at index {index} can not null. Meaning: số lượng tại 1 không thể rỗng \n Message: quantity at index {index} can less one. Meaning: số lượng tại 1 không thể nhỏ hơn 1 \n status_code : 999 \n Message: system error. Meaning: Lỗi hệ thống , liên hệ admin.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number"
                    },
                    "address": {
                      "type": "string"
                    },
                    "createdDate": {
                      "type": "string"
                    },
                    "updatedDate": {
                      "type": "string"
                    },
                    "beginDate": {
                      "type": "string"
                    },
                    "endDate": {
                      "type": "string"
                    },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "groupType": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        },
                        "createdDate": {
                          "type": "string"
                        },
                        "additionalConfig": {
                          "type": "object",
                          "properties": {
                            "cleanUpPoint": {
                              "type": "number"
                            },
                            "buyingStuffPoint": {
                              "type": "number"
                            },
                            "supportConferencePoint": {
                              "type": "number"
                            },
                            "trainingPoint": {
                              "type": "number"
                            },
                            "weeklyCleanUpPoint": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    },
                    "eventUserList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user": {
                            "type": "object",
                            "properties": {
                              "user": {
                                "type": "object",
                                "properties": {
                                  "username": {
                                    "type": "string"
                                  },
                                  "lastName": {
                                    "type": "string"
                                  },
                                  "firstName": {
                                    "type": "string"
                                  },
                                  "fullName": {
                                    "type": "string"
                                  },
                                  "email": {
                                    "type": "string"
                                  },
                                  "userRole": {
                                    "type": "array",
                                    "items": {
                                      "type": "string"
                                    }
                                  },
                                  "avatar": {
                                    "type": "string"
                                  },
                                  "birthday": {
                                    "type": "date"
                                  },
                                  "numberPhone": {
                                    "type": "string"
                                  },
                                  "gmail": {
                                    "type": "string"
                                  },
                                  "skype": {
                                    "type": "string"
                                  },
                                  "dateStartWork": {
                                    "type": "string"
                                  }
                                }
                              },
                              "event": {
                                "type": "object"
                              },
                              "type": {
                                "type": "number"
                              }
                            }
                          },
                          "event": {
                            "type": "object"
                          },
                          "type": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "errorCode": {
                        "type": "number"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "example": {
                "data": {
                  "id": 32,
                  "name": "Support",
                  "description": null,
                  "status": 1,
                  "address": null,
                  "createdDate": "10-09-2018 15:36",
                  "updatedDate": "10-09-2018 15:36",
                  "beginDate": "10-09-2018 11:00",
                  "endDate": null,
                  "group": {
                    "id": 2,
                    "name": "Support",
                    "description": null,
                    "groupType": {
                      "id": 4,
                      "name": "Support chung"
                    },
                    "createdDate": "07-09-2018 18:02",
                    "additionalConfig": {
                      "cleanUpPoint": 1,
                      "buyingStuffPoint": 2,
                      "supportConferencePoint": 3,
                      "trainingPoint": 4,
                      "weeklyCleanUpPoint": 5
                    }
                  },
                  "eventUserList": [
                    {
                      "user": {
                        "username": "kpi.admin",
                        "lastName": "kpi",
                        "firstName": "admin",
                        "fullName": null,
                        "email": "kpi.admin@higgsup.com",
                        "userRole": null,
                        "avatar": "1",
                        "birthday": "2018-09-25",
                        "numberPhone": "0966211618",
                        "address": "Lê văn Thành ",
                        "gmail": "lethanh9398@gmail.com",
                        "skype": "solitarysp@gmail.com",
                        "dateStartWork": "2018-09-29"
                      },
                      "type": 2
                    }
                  ],
                  "additionalConfig": [
                    {
                      "name": "cleanUpPoint",
                      "point": null,
                      "quantity": 1
                    }
                  ]
                },
                "status_code": 200
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/groups/club": {
      "post": {
        "tags": [
          "group-controller"
        ],
        "summary": "Create a club activity",
        "operationId": "createClubUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "groupDTO",
            "description": "groupDTO",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "groupType",
                "additionalConfig"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "groupType": {
                  "required": [
                    "id"
                  ],
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                },
                "additionalConfig": {
                  "type": "object",
                  "required": [
                    "host",
                    "minNumberOfSessions",
                    "participationPoint",
                    "effectivePoint"
                  ],
                  "properties": {
                    "host": {
                      "type": "string"
                    },
                    "minNumberOfSessions": {
                      "type": "number"
                    },
                    "participationPoint": {
                      "type": "number"
                    },
                    "effectivePoint": {
                      "type": "number"
                    }
                  }
                }
              },
              "example": {
                "name": "Câu lạc bố Đá Bóng",
                "groupType": {
                  "id": 2
                },
                "additionalConfig": {
                  "host": "kpi.man",
                  "minNumberOfSessions": 2,
                  "participationPoint": 3,
                  "effectivePoint": 4
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "status_code : 200. Meaning: Thành công \n status_code : 904. Message: json processing exception. Meaning: Lỗi convert json \n status_code : 903. Message: parameter name is not valid. Meaning: Tên câu lạc bộ không hợp lệ \n status_code : 901. Message: parameter min number of sessions is not valid. Meaning: Số lần tổi thiểu đăng ký không hợp lệ \n status_code : 901. Message: parameter point is not valid. Meaning: Điểm không hợp lệ \n status_code : 901. Message: group type does not exist. Meaning: Loại group không tồn tại \n  status_code : 932. Message: group name already exist. Meaning:  Tên group đã tồn tại \n status_code : 914. Message: minimum session must be bigger than zero. Meaning:  Số buổi tối thiểu phải lớn hơn 0 ",
            "schema": {
              "type": "object",
              "required": [
                "id",
                "name",
                "groupType",
                "additionalConfig"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "groupType": {
                  "required": [
                    "id"
                  ],
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    }
                  }
                },
                "createdDate": {
                  "type": "string"
                },
                "additionalConfig": {
                  "type": "object",
                  "required": [
                    "host",
                    "minNumberOfSessions",
                    "participationPoint",
                    "effectivePoint"
                  ],
                  "properties": {
                    "host": {
                      "type": "string"
                    },
                    "minNumberOfSessions": {
                      "type": "number"
                    },
                    "participationPoint": {
                      "type": "number"
                    },
                    "effectivePoint": {
                      "type": "number"
                    }
                  }
                }
              },
              "example": {
                "data": {
                  "id": 10,
                  "name": "Câu lạc bố Đá Bóng",
                  "description": null,
                  "groupType": {
                    "id": 2
                  },
                  "createdDate": "07-09-2018 11:35",
                  "additionalConfig": {
                    "host": "kpi.man",
                    "minNumberOfSessions": 2,
                    "participationPoint": 3,
                    "effectivePoint": 4
                  },
                  "status_code": 200
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
...
}